pyproject.toml
[tool.poetry]
name = "f451-sensors"
version = "0.0.0"
description = "f451 Sensors"
authors = ["Martin Lanser <martinlanser@gmail.com>"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/mlanser/f451-sensors"
repository = "https://github.com/mlanser/f451-sensors"
documentation = "https://f451-sensors.readthedocs.io"
classifiers = [
    "Development Status :: 1 - Planning",
]

[tool.poetry.urls]
Changelog = "https://github.com/mlanser/f451-sensors/releases"

[tool.poetry.dependencies]
python = "^3.8"
paho-mqtt = "^1.6.1"
Faker = "^13.13.0"
influxdb-client = {extras = ["ciso"], version = "^1.29.1"}
argparse = "^1.4.0"
rich = "^12.4.4"
konsole = "^0.7.0"

[tool.poetry.dev-dependencies]
Pygments = ">=2.10.0"
black = ">=21.10b0"
coverage = {extras = ["toml"], version = ">=6.2"}
darglint = ">=1.8.1"
flake8 = ">=4.0.1"
flake8-bandit = ">=2.1.2"
flake8-bugbear = ">=21.9.2"
flake8-docstrings = ">=1.6.0"
flake8-rst-docstrings = ">=0.2.5"
furo = ">=2021.11.12"
isort = ">=5.10.1"
mypy = ">=0.930"
pep8-naming = ">=0.12.1"
pre-commit = ">=2.16.0"
pre-commit-hooks = ">=4.1.0"
pytest = "^7.0.1"
pytest-mock = "^3.7.0"
pyupgrade = ">=2.29.1"
safety = ">=1.10.3"
sphinx = ">=4.5.0"
sphinx-autobuild = ">=2021.3.14"
sphinx-click = ">=4.0.0"
sphinx-rtd-theme = "^1.0.0"
sphinx-argparse-cli = "^1.8.3"
typeguard = ">=2.13.3"
xdoctest = {extras = ["colors"], version = ">=0.15.10"}
myst-parser = {version = ">=0.16.1"}

[tool.poetry.scripts]
f451-sensors = "f451_sensors.__main__:main"

[tool.pytest.ini_options]
testpaths = [
    "tests"
]
markers = [
    "smoke",    # quick smoke test with 3rd-party component mocked
    "slow",     # mark test as slow (deselect with '-m "not slow"')
]

[tool.coverage.paths]
source = ["src", "*/site-packages"]
tests = ["tests", "*/tests"]

[tool.coverage.run]
branch = true
source = ["f451_sensors", "tests"]

[tool.coverage.report]
show_missing = true
fail_under = 100

[tool.isort]
profile = "black"
force_single_line = true
lines_after_imports = 2

[tool.mypy]
strict = true
warn_unreachable = true
pretty = true
show_column_numbers = true
show_error_codes = true
show_error_context = true
exclude = [
    '_temp_.*\.py',     # Don't run mypy on temp files
    'temp.*\.py',
    'conftest\.py',     # Don't run mypy on test files
    'test_.*\.py',
]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
